refactor(doc/codesandbox): renamed file constants to code
