+++ packages/doc/content/components/list/web.mdx
@@ -1,5 +1,3 @@
-import ExpoSnack from 'react-expo-snack';
-
 ### Reference
 
 Gympass `<List />` component implements all our design guidelines, be sure to wrap your items with `<List.Item/>` that uses default spacing, border and color pallete tokens.
@@ -10,4 +8,4 @@ Gympass `<List />` component implements all our design guidelines, be sure to wr
 
 ### Usage
 
-<CodeSandbox>{`<Button>Test</Button>`}</CodeSandbox>
+<CodeSandbox>{'<Button>Alo alyso</Button>'}</CodeSandbox>
